+++ src/judging/text.rs
@@ -5,7 +5,6 @@ use crate::testsuite::FloatErrors;
 use crossbeam_utils::atomic::AtomicCell;
 use itertools::{EitherOrBoth, Itertools as _};
 use once_cell::sync::Lazy;
-use once_cell::sync_lazy;
 use serde::{Serialize, Serializer};
 use serde_derive::Serialize;
 use smallvec::SmallVec;
@@ -62,8 +61,8 @@ impl Text {
                     wtr.write_str(s)?;
                     wtr.reset()?;
                 }
-                Word::U0020s(n) => wtr.write_spaces(n.get())?,
-                Word::TrailingU0020s(n) => {
+                Word::Sps(n) => wtr.write_spaces(n.get())?,
+                Word::TrailingSps(n) => {
                     wtr.set_color(color!(bg(Yellow), intense, bold))?;
                     wtr.write_spaces(n.get())?;
                     wtr.reset()?;
@@ -132,10 +131,10 @@ impl TextDiff {
                                     } = FLOAT_ERRORS.load();
                                     (v1 - v2).abs() <= d || ((v1 - v2) / v2).abs() <= r
                                 }
-                                (Word::U0020s(n1), Word::U0020s(n2))
-                                | (Word::U0020s(n1), Word::TrailingU0020s(n2))
-                                | (Word::TrailingU0020s(n1), Word::U0020s(n2))
-                                | (Word::TrailingU0020s(n1), Word::TrailingU0020s(n2)) => n1 == n2,
+                                (Word::Sps(n1), Word::Sps(n2))
+                                | (Word::Sps(n1), Word::TrailingSps(n2))
+                                | (Word::TrailingSps(n1), Word::Sps(n2))
+                                | (Word::TrailingSps(n1), Word::TrailingSps(n2)) => n1 == n2,
                                 (Word::Plain(s1), Word::Plain(s2))
                                 | (Word::UnicodeEscape(s1), Word::UnicodeEscape(s2)) => s1 == s2,
                                 (Word::Tab, Word::Tab)
@@ -161,7 +160,7 @@ impl TextDiff {
         }
 
         static FLOAT_ERRORS: Lazy<AtomicCell<FloatErrorsNormalized>> =
-            sync_lazy!(AtomicCell::new(FloatErrorsNormalized::default()));
+            Lazy::new(|| AtomicCell::new(FloatErrorsNormalized::default()));
 
         if let Some(float_errors) = float_errors {
             FLOAT_ERRORS.store(FloatErrorsNormalized {
@@ -272,8 +271,8 @@ impl TextDiff {
                         wtr.write_str(s)?;
                         wtr.reset()?;
                     }
-                    Word::U0020s(n) => wtr.write_spaces(n.get())?,
-                    Word::TrailingU0020s(n) => {
+                    Word::Sps(n) => wtr.write_spaces(n.get())?,
+                    Word::TrailingSps(n) => {
                         wtr.set_color(color!(bg(Yellow), intense, bold))?;
                         wtr.write_spaces(n.get())?;
                         wtr.reset()?;
@@ -319,11 +318,11 @@ impl TextDiff {
                         wtr.set_color(color!(fg(Red), intense, underline))?;
                         wtr.write_str(s)?;
                     }
-                    Word::U0020s(n) => {
+                    Word::Sps(n) => {
                         wtr.set_color(color!(fg(Red), intense, underline))?;
                         wtr.write_spaces(n.get())?
                     }
-                    Word::TrailingU0020s(n) => {
+                    Word::TrailingSps(n) => {
                         wtr.set_color(color!(fg(Red), bg(Yellow), intense, underline))?;
                         wtr.write_spaces(n.get())?
                     }
@@ -479,11 +478,11 @@ fn parse_words(s: &str, floats: bool) -> impl Iterator<Item = Word> {
         let tab = recognize(char('\t')).map(|_| Word::Tab);
         let cr = recognize(char('\r')).map(|_| Word::Cr);
         let lf = recognize(char('\n')).map(|_| Word::Lf);
-        let codepoints = recognize(skip_many1(satisfy(|c: char| {
+        let unicode_escape = recognize(skip_many1(satisfy(|c: char| {
             (c.is_whitespace() || c.is_control()) && ![' ', '\t', '\r', '\n'].contains(&c)
         })))
         .map(Word::UnicodeEscape);
-        choice((plain_or_float, tab, cr, lf, codepoints))
+        choice((plain_or_float, tab, cr, lf, unicode_escape))
     }
 
     let with_u0020s = recognize(skip_many1(char(' ')))
@@ -491,9 +490,9 @@ fn parse_words(s: &str, floats: bool) -> impl Iterator<Item = Word> {
         .map::<_, SmallVec<[Word; 2]>>(|(s, word): (&str, _)| {
             let n = NonZeroUsize::new(s.len()).unwrap();
             match word {
-                Some(Word::Lf) => [Word::TrailingU0020s(n), Word::Lf].into(),
-                Some(word) => [Word::U0020s(n), word].into(),
-                None => iter::once(Word::TrailingU0020s(n)).collect(),
+                Some(Word::Lf) => [Word::TrailingSps(n), Word::Lf].into(),
+                Some(word) => [Word::Sps(n), word].into(),
+                None => iter::once(Word::TrailingSps(n)).collect(),
             }
         });
 
@@ -503,7 +502,7 @@ fn parse_words(s: &str, floats: bool) -> impl Iterator<Item = Word> {
     ))
     .skip(eof())
     .easy_parse(State::with_positioner(s, IndexPositioner::new()))
-    .unwrap()
+    .unwrap_or_else(|e| unreachable!("{:?}", e))
     .0
     .into_iter()
     .flatten()
@@ -512,8 +511,8 @@ fn parse_words(s: &str, floats: bool) -> impl Iterator<Item = Word> {
 #[derive(Debug, Clone, Copy, Serialize)]
 enum Word<'a> {
     Plain(&'a str),
-    U0020s(NonZeroUsize),
-    TrailingU0020s(NonZeroUsize),
+    Sps(NonZeroUsize),
+    TrailingSps(NonZeroUsize),
     Tab,
     Cr,
     Lf,
@@ -532,7 +531,7 @@ impl Word<'_> {
     fn width(self, str_width: fn(&str) -> usize) -> Option<NonZeroUsize> {
         match self {
             Word::Plain(s) => NonZeroUsize::new(str_width(s)),
-            Word::U0020s(n) | Word::TrailingU0020s(n) => Some(n),
+            Word::Sps(n) | Word::TrailingSps(n) => Some(n),
             Word::Tab | Word::Cr => NonZeroUsize::new(2),
             Word::UnicodeEscape(s) => NonZeroUsize::new(8 * s.chars().count()),
             Word::Float { string, .. } => NonZeroUsize::new(str_width(string)),
@@ -660,7 +659,6 @@ mod tests {
 
     use failure::Fallible;
     use once_cell::sync::Lazy;
-    use once_cell::sync_lazy;
     use pretty_assertions::assert_eq;
     use stable_deref_trait::StableDeref;
     use termcolor::{Ansi, Color, ColorSpec, WriteColor};
@@ -687,17 +685,17 @@ mod tests {
         static EXPECTED2: &str = "42.0\n";
 
         static INPUT3: &str = "foo\nbar";
-        static EXPECTED3: Lazy<String> = sync_lazy! {
+        static EXPECTED3: Lazy<String> = Lazy::new(|| {
             let mut wtr = Ansi::new(vec![]);
             wtr.write_all(b"foo\nbar").unwrap();
             wtr.set_color(&fg_yellow_intense_bold()).unwrap();
             wtr.write_all(b"<noeol>\n").unwrap();
             wtr.reset().unwrap();
             String::from_utf8(wtr.into_inner()).unwrap()
-        };
+        });
 
         static INPUT4: &str = "foo ";
-        static EXPECTED4: Lazy<String> = sync_lazy! {
+        static EXPECTED4: Lazy<String> = Lazy::new(|| {
             let mut wtr = Ansi::new(vec![]);
             wtr.write_all(b"foo").unwrap();
             wtr.set_color(&bg_yellow_intense_bold()).unwrap();
@@ -707,10 +705,10 @@ mod tests {
             wtr.write_all(b"<noeol>\n").unwrap();
             wtr.reset().unwrap();
             String::from_utf8(wtr.into_inner()).unwrap()
-        };
+        });
 
         static INPUT5: &str = "[ \x1b ]\n";
-        static EXPECTED5: Lazy<String> = sync_lazy! {
+        static EXPECTED5: Lazy<String> = Lazy::new(|| {
             let mut wtr = Ansi::new(vec![]);
             wtr.write_all(b"[ ").unwrap();
             wtr.set_color(&fg_yellow_intense_bold()).unwrap();
@@ -718,7 +716,7 @@ mod tests {
             wtr.reset().unwrap();
             wtr.write_all(b" ]\n").unwrap();
             String::from_utf8(wtr.into_inner()).unwrap()
-        };
+        });
 
         test(INPUT1, EXPECTED1)?;
         test(INPUT2, &EXPECTED2)?;
@@ -740,7 +738,7 @@ mod tests {
 
         static LEFT: &str = "foo\n";
         static RIGHT: &str = "foo";
-        static EXPECTED: Lazy<String> = sync_lazy! {
+        static EXPECTED: Lazy<String> = Lazy::new(|| {
             let mut wtr = Ansi::new(vec![]);
             wtr.write_all("│".as_ref()).unwrap();
             wtr.set_color(&fg_magenta_bold()).unwrap();
@@ -762,7 +760,7 @@ mod tests {
             wtr.reset().unwrap();
             wtr.write_all("│\n".as_ref()).unwrap();
             String::from_utf8(wtr.into_inner()).unwrap()
-        };
+        });
 
         test(LEFT, RIGHT, &EXPECTED)
     }