+++ src/macros.rs
@@ -135,7 +135,7 @@ macro_rules! plural {
 macro_rules! lazy_regex {
     ($expr:expr) => {{
         static REGEX: ::once_cell::sync::Lazy<::regex::Regex> =
-            ::once_cell::sync_lazy!(::regex::Regex::new($expr).unwrap());
+            ::once_cell::sync::Lazy::new(|| ::regex::Regex::new($expr).unwrap());
         &REGEX
     }};
     ($expr:expr,) => {
@@ -146,7 +146,9 @@ macro_rules! lazy_regex {
 macro_rules! selector {
     ($selectors:literal) => {{
         static SELECTOR: ::once_cell::sync::Lazy<::scraper::selector::Selector> =
-            ::once_cell::sync_lazy!(::scraper::selector::Selector::parse($selectors).unwrap());
+            ::once_cell::sync::Lazy::new(|| {
+                ::scraper::selector::Selector::parse($selectors).unwrap()
+            });
         &SELECTOR
     }};
     ($selectors:literal,) => {